Completed Refactoring
