--- script/button.css
@@ -1,25 +1,0 @@
-/* Define the button style */
-button {
-    background-color: #007acc;
-    /* VSCode blue */
-    color: rgb(255, 255, 255);
-    /* Text color */
-    border: none;
-    /* Remove default border */
-    padding: 10px 20px;
-    /* Padding for the button */
-    border-radius: 4px;
-    /* Rounded corners */
-    font-size: 14px;
-    /* Font size */
-    cursor: pointer;
-    /* Pointer cursor on hover */
-    transition: background-color 0.3s;
-    /* Smooth transition for hover effect */
-}
-
-/* Hover effect */
-button:hover {
-    background-color: #005a9e;
-    /* Darker shade of blue for hover */
-}
