--- script/constant.js
@@ -1,7 +1,0 @@
-const ELEMENT_IDS = {
-    TRIGGER_MESSAGE_BUTTON: 'trigger-show-message-button'
-};
-
-const POST_MESSAGE_ACTION = {
-    SHOW_WARNING_LOG: 'SHOW_WARNING_LOG'
-}
